chore: remove duplicate in browser toolkit
